+++ ddth-commons-core/src/test/java/com/github/ddth/commons/qnd/rocksdb/QndRocksDbWrapper.java
@@ -1,31 +1,3 @@
-# com.github.ddth.commons.rocksdb
-
-Utility and helper classes to work with [RocksDB](http://rocksdb.org).
-
-_**Available since v0.8.0.**_
-
-## Maven
-
-```xml
-<dependency>
-    <groupId>com.github.ddth</groupId>
-    <artifactId>ddth-commons-rocksdb</artifactId>
-    <version>${ddth_commons_version}</version>
-    <type>pom</type>
-</dependency>
-```
-
-**Class `RocksDbUtils`**
-
-- Helper methods to open RocksDB database.
-- Wrapper class to GET/PUT/DELETE RocksDB data.
-
-
-**Class `RocksDbWrapper`**
-
-## Examples
-
-```java
 package com.github.ddth.commons.qnd.rocksdb;
 
 import java.io.File;
@@ -64,12 +36,3 @@ public class QndRocksDbWrapper {
         }
     }
 }
-```
-
-## History
-
-**v0.9.4 - 2019-04-29**
-- Upgrade to `rocksdbjni:6.0.1`.
-
-**v0.8.0 - 2018-01-22**
-- First release